Add Text to 'To Do' List
I added a click event so that I can submit a new text to add to my to do lists. I use delegation and target all three buttons.
If the input is empty an alert message appears saying that is not allowed. I clear the input after every submit. I create an li element to add to my page containing the input text on submit.