Local Storage
I made it so local Storage is fully functional and passes the tests. Every time the page reloads, if there is localStorage, I turn it into an object and loop through it, and add all the elements back  on the page. Every time I add an element I update the listStorage object and I update the local Storage.