Double Click Event
First, I added a click event and delegated it to all the 'li' elements in tasksContainer.

Second, I made so on the double click, the list element is replaced with a input element. Focus is immediately placed on the element.

Third, I add an event for when the focus is taken off the input ('blur'). When this event happens I update the list element and replace the input with the  list element.

Fourth, I update the local storage, by finding the index of the list element in the DOM(in 'ul') and then use that index in listStorage, to change the content there. I decided with using this method instead of directly finding the index in the listStorage because there could theoretically be two of the same tasks, and if I use index of directly in the listStorage it would alter the first one automatically.

PS. on line 76 I also make sure you can't leave it blank.