Fixed Bug and Added Local Storage
There was a bug where the element I moved would still move after I wasn't hovering over it. so When I moved out of the element I made the value of the element being hovered over null. This way there is no value for the key command to move.

I added local storage to my alt command. So it can be saved.