Search Filter
I add an input event to the element with the id search. The search filter function attached to the event hides all the elements that do not include the text in the search box. I also unhide all the elements so that the elements do not stay hidden forever after each change in text, rather what elements are hidden and not hidden is refreshed on each change in text.