CSS Styling
I Completed my CSS styling adding a wood theme to my page. I changed everything from the color of the selection to the styling of the scroll bar. I changed the way the pointer looks hovering over the elements on the page, and made the styling change when list elements and buttons were being hovered over. I made important elements "sticky" so they are never out of site of the user. I payed really close attention to detail, in an effort to give the page an immersive feel.