Adding Tasks With the Enter Key
I add a click event listener which gets the input field the user is focused on and his adjacent add button. I add a letdown event listener which on 'Enter 'clicks the add button for the user.  I make it so after the add button is clicked the focus is brought directly on the input field again. This way the user can continuously add tasks without picking up there hand and clicking with the mouse.